trekk funksjon for å hente fagsaker på person ut i egen api-hook så OppgaverContext ikke trenger tilgang til FagsakContext
